New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): pipeline            Description: Describe your changes here...
Bug fix on baseline subtraction. Support the case when there are 
no update from previous iteration in line detection and validation
so that blfile doesn't exist. Just skip updating baseline table.  

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@25506 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ pipeline/pipeline/hsd/tasks/baseline/fitting.py
@@ -175,10 +175,11 @@ class FittingBase(object):
                     tb.putcell('SPECTRA', rows[i], spectra[i])
 
             # update baseline table
-            self._update_bltable(bltable_name, blfile, datatable, index_list, nchan, _edge, nwindow, fragment)
+            if os.path.exists(blfile):
+                self._update_bltable(bltable_name, blfile, datatable, index_list, nchan, _edge, nwindow, fragment)
 
-            # cleanup blfile
-            os.system('rm -rf %s'%(blfile))
+                # cleanup blfile
+                os.system('rm -rf %s'%(blfile))
                         
     def _calc_baseline_fit(self, scan, data, polyorder, nchan, modification, edge, masklist, blfile, win_polyorder, fragment, nwindow, mask):
         LOG.debug('masklist=%s'%(masklist))
